+++ packages/paths/src/helpers/linear.ts
@@ -1,3 +1,4 @@
+// Copied from: https://github.com/rveciana/svg-path-properties
 import type {Point, PointProperties} from './types';
 
 export const makeLinearPosition = (
