+++ packages/paths/src/index.ts
@@ -1,0 +1,1 @@
+export {getLength} from './get-length';
